requirements.txt
streamlit==1.40.2
google-generativeai==0.8.3
pillow==11.0.0
requests==2.32.3
python-dotenv==1.0.1
